Update file-lock dependency to v2.1.4
This fixes a security vulnerability in the `nix` dependency of
`file-lock`, which affects `nix` < 0.20.2.

From v1.1 to v2.1 of `file-lock`, the `FileLock::lock()` function
changed:  As third argument, it now takes a `FileOptions` object instead
of a `bool` to indicate if the lock should be writable.  Internally,
`file-lock` uses `std::fs::OpenOptions`, and `FileOptions` is just a
wrapper around the former.  This commit constructs a `FileOptions`
object with the same attributes as the corresponding `OpenOptions` had
before in the `file-lock` implementation.  Thus, even though it may seem
weird to get the *read* permission *only* for read-only locks, this
matches the prior implementation.